+++ test/convert.test.js
@@ -1,5 +1,5 @@
-const { convert } = require('../');
-const expect = require('chai').expect;
+import { convert } from '../lib/index';
+import { expect, describe, it } from 'vitest';
 
 describe('convert', () => {
   it('[convert]default', () => {
@@ -56,5 +56,15 @@ describe('convert', () => {
   it('[convert]numToSymbol abnormal', () => {
     const result = convert('l2', { format: 'numToSymbol' });
     expect(result).to.be.equal('l2');
+  });
+
+  it('[convert]numToSymbol iu', () => {
+    const result = convert('liu2', { format: 'numToSymbol' });
+    expect(result).to.be.equal('liú');
+  });
+
+  it('[convert]special tone', () => {
+    const result = convert('m̄ hm ê̄ ế ê̌ ề', { format: 'symbolToNum' });
+    expect(result).to.be.equal('m1 hm0 ê1 ê2 ê3 ê4');
   });
 });