+++ types/common/constant.d.ts
@@ -1,6 +1,7 @@
 export declare const DoubleUnicodePrefixReg: RegExp;
 export declare const DoubleUnicodeSuffixReg: RegExp;
 export declare const DoubleUnicodeReg: RegExp;
+export declare const DoubleUnicodeCharReg: RegExp;
 export declare const enum Probability {
     Unknown = 1e-13,
     Rule = 1e-12,
@@ -8,3 +9,8 @@ export declare const enum Probability {
     Surname = 1,
     Custom = 1
 }
+export declare const Priority: {
+    Normal: number;
+    Surname: number;
+    Custom: number;
+};