+++ types/core/pinyin/middlewares.d.ts
@@ -1,5 +1,5 @@
-import type { SingleWordResult } from '../../common/type';
-import { CompleteOptions } from './index';
+import type { SingleWordResult } from "../../common/type";
+import { CompleteOptions } from "./index";
 export declare const validateType: (word: unknown) => boolean;
 export declare const middleWareNonZh: (list: SingleWordResult[], options: CompleteOptions) => SingleWordResult[];
 export declare const middlewareMultiple: (word: string, options: CompleteOptions) => SingleWordResult[] | false;
@@ -20,5 +20,4 @@ export declare const middlewareType: (list: SingleWordResult[], options: Complet
     polyphonic: string[];
     inZhRange: boolean;
 }[];
-export declare const middlewareDoubleUnicode: (list: SingleWordResult[]) => SingleWordResult[];
 export declare const middlewareToneSandhi: (list: SingleWordResult[], toneSandhi: boolean) => SingleWordResult[];